Strix 2.1.2
Feats
1. `PopularityUtil` 更新.
2. `RedisUtil` 更新. 增加 `ZSET` 相关方法, 增加 `isType` 方法.

Fixes

+++ src/main/java/cn/projectan/strix/model/db/PopularityData.java
@@ -34,7 +34,7 @@ public class PopularityData extends BaseModel {
 
     private String dataId;
 
-    private Integer originalValue;
+    private Long originalValue;
 
     public PopularityData(String createBy, String updateBy) {
         super(createBy, updateBy);
